Rejiggering into a simpler architecture.
At the core, the message broker now both dispatches events related
to message traffic on the stream and it allows us to install
hanlders for these messages. The message handlers are specific to
each message type, simplifying the dispatch logic. Installing and
clearing the handlers now happens directly from FEBEConnection.
This means that FEBEConnection is a little bulky, but it needs
to maintain a lot of state across these different message handlers;
there is no simple way to get around this.

This also introduces some changes to the codec package: encoding
and decoding are now separate, but different decoding formats
are done together. This keeps the decoded management a little
more sane, although we could probably use an abstract base class
to unify some of the repetitive codec work.

Also some more hoisting of constants to static classes, and some
event rejiggering. We may want to move some of this out of the
febe package.