A rather different take on the API.
This is starting to feel like one-man bikeshedding, but I'm significantly
happier with this approach. The basic idea is to do away with the
event dispatch method (under the assumption that a query will typically
have a single consumer), and simply register handlers with each submitted
query. The handler will get callbacks when the query produces results.
This eleminates a number of extraneous classes and other API ugliness.