Tweaking test base to connect once per test class (rather than once per test method).
We make the assumption that the connection stays in a
valid state throughout all the test methods in the test
class. If there are some negative tests that require this
not to be true, the tests can create separate connections
through `createConnection()`. Note that this method is
asynchronous and should be given enough time to complete
with the `async` metadata parameter on the calling method.