Rethinking the IDataStream API.
No more connected / disconnected. Just write until you hit an error. Client
is always in charge of closing the stream, and always assumes that it is closed
successfully. In theory, this may be an undesirable simplification in some
niche cases, but in practice, this will make the protocol model much simpler
with no real loss in flexibility.

Also reworking the ByteArrayStream class into a proper IDataStream.