Adding cancel() to IConnection interface; tweaking implementation.
Cancel now returns true if a successful query cancel attempt was
made. A successful attempt means either a pending statement was
dequeued for execution, or the statement to be cancelled is
currently executing and a Cancel message was sent to the
backend. Note that there is an inherent race condition in
cancelling running queries, so 'a successful query cancel
attempt' does not necessarily mean the query will be cancelled.