Rejiggering codec mechanism.
Previously, the idea was that the same decoders and encoders could
be used to produce different encoded/decoded types, and this was
baked into the CodecFactory. In reality, that's probably overkill:
if a decoder/encoder will produce different output types, the only
thing which should affect that is the type information of the data
to be encoded / decoded. This has now been folded into the encoder
and decoder interfaces themselves, giving more flexibility there
while simplifying the CodecFactory.