Restoring functionality to type-guess 'int' for integral numbers.
This affects uint, int, and Number (basically, anything that "fits"
in an int will be considered an int), but that's probably more useful
than automatically "promoting" ints to double precision floats.

We can provide wrappers at a later date to allow integral Numbers
to be passed in as double precision.