Slight change in result handler base class semantics.
Previously, the handleCompletion hook would be overridden fully. Now, we
introduce a separate function to be overridden, and manage cleanup of the
previous result set (if any) in the main event handler. This makes the
behavior a little more sane in the case where a single (simple) query
has multiple completions (due to multiple component queries), but only
some of the actually return results. Previously, it was really easy to
lose track of which result set belonged to whom; this helps that case.