Deleting outdated directories.
Pegasus-fx is probably not a good idead in the first place--in all but the tiniest
of apps, it will make sense to set up a Service in some sort of MVC architecture
and not do anything too clever. Pgfx-console has turned into pgconsole. The harness
was really an early version of pgconsole, and between that and the functional test,
it is not needed.